requirements.txt
fastapi>=0.95
uvicorn[standard]>=0.20
google-genai>=0.3.0
python-multipart>=0.0.5
python-docx>=0.8.11
python-pptx>=0.6.21
Pillow>=9.0.0
pytesseract>=0.3.10
pydub>=0.25.1
whisper>=20230314
python-dotenv>=1.0.0
pytest>=7.0.0
requests>=2.28.0